software/requirements.txt
paho-mqtt
discord.py
asyncio
aihttp
python-dotenv
numpy
